# Benchmark and fixes