--- a/news/477.feature
+++ b/news/477.feature
@@ -1,0 +1,1 @@
+Optimized config creation time