Closes #267
Closes #477

1. Adds a new benchmark using [pytest-benchmark](https://pytest-benchmark.readthedocs.io/en/latest/index.html).
Currently it tests creation from dict and from DictConfig. Notably creation from DictConfig is 5 times slower and needs some work.

The benchmark is executed in nox.
To run it manually:
```
$ pytest benchmark/benchmark.py
```

2. Fixed a several bugs with shallow copy
3. Added some more tests ensuring deepcopy semantics.

Optimizations:

1. Added flags cache. This improves initialization time from a large dict in the benchmark from 600ms to 450ms.
It also slows down the already slow the initialization from a large DictConfig from 1300ms to 1500ms, but this one needs more work anyway.

2. Optimized deepcopy for DictConfig (creation speedup from a large DictConfig is about 5x).
3. Optimized deepcopy for List (creation speedup from a large DictConfig containing ListConfig leaves is about 5x).