@omry 2021-01-21
Comparing results before and after the optimizations seems too good to be true, especially for the large dict config.
```
$ pytest benchmark/benchmark.py --benchmark-autosave --benchmark-compare 0023 --benchmark-sort=name
```
![image](https://user-images.githubusercontent.com/376455/105269601-18198d00-5b49-11eb-85ae-2f3da2bbcf14.png)