@omry 2021-01-21
Usually when something is too good to be true, it's actually is.
By switching to shallow copy, the semantics of initialization from DictConfig became different than the semantics of initializing from dict.
It was so much faster because it didn't copy, this actually broke the source config simply because a node can't have two parents.
This simply happened one level deeper than the level tested in the previous tests.

Once I realized that it could be the case, I added another test that proved it (050bbaaa4d749b0646dda3c3be2b2d03460728a3).

Then finally spent some time trying to understand why deepcopy was so god damned slow.
The implementation of deepcopy for nodes is deepcopying the parent node, which means we are copying the entire config tree many more times than  we need to.

I implemented a more custom version of `__deepcopy__` that  side stepped the issue by assigning None as the parent before copying, deepcopying and then restoring the parent.

The benchmark results makes more sense now.
Initializing from both dict and DictConfig takes about the same time (and both faster than before).

![image](https://user-images.githubusercontent.com/376455/105320589-326c5e80-5b7b-11eb-92e2-eb0feb3712d7.png)