@omry 2021-01-21
Reorganized commits and updated description.
This has became a pretty large diff. I am going to land it now, but please feel free to review and comment.